refactor: 优化空字符串检查及代码格式
- 将`isNullOrBlank()`替换为更高效的`isNotBlank()`
- 移除未使用的`Locale`导入并调整导入顺序
- 简化`formatDuration`函数的条件判断
- 统一颜色常量的表示方式

+++ app/src/main/java/com/xzyht/notifyrelay/feature/notification/superisland/floating/SmallIsland/compose/BCompose.kt
@@ -206,7 +206,7 @@ fun BCompose(
                     CircularProgressCompose(
                         progress = bComp.progress,
                         colorReach = Color(parseColorSafe(bComp.colorReach, 0xFF3482FF.toInt())),
-                        colorUnReach = Color(parseColorSafe(bComp.colorUnReach, 0x33333333.toInt())),
+                        colorUnReach = Color(parseColorSafe(bComp.colorUnReach, 0x33333333)),
                         strokeWidth = 2.5.dp,
                         isClockwise = !bComp.isCCW,
                         size = size
